Create --usr-json-path argument
This creates a new setting for unFTP that allows you to point it to a
JSON file that will contain user specific settings. This may look like
this:

```
[
  {
    "username": "alice",
    "password": "alice",
    "name": "Alice",
    "surname": "From Wonderland",
    "vfs_perms": ["-mkdir","-rmdir","-del","-ren", "-md5"],
    "allowed_mime_types": ["text/plain"],
    "root": "alice",
    "account_enabled": true,
    "some_other": "value"
  },
  {
    "username": "bob",
    "password": "bob",
    "name": "Bob",
    "surname": "The Builder",
    "vfs_perms": ["none", "+put", "+md5"],
    "allowed_mime_types": ["text/plain"],
    "root": "bob",
    "account_enabled": true,
    "some_other": "value"
  }
]
```

The "allowed_mime_types" and "root" is not implemented yet, "some_other"
shows that it ignores settings it doesn't know about and "password"
belong to the JSON authenticator. We are using the same file for both
--auth-json-path and --usr-json-path.

The "vfs_perms" and "account_enabled" properties does work.