Implement per-user root directory
This builds on github PR #97 to add a configurable optional home
directory per user as requested in issue #85.

The user's home gets specified in a JSON file pointed to by the
`---usr-json-path` argument. The `root` property needs to be set:

```json
[
  {
    "username": "alice",
    "password": "12345678",
    "root": "alice"
  },
  {
    "username": "bob",
    "password": "secret",
    "root": "bob"
  }
]
```

At the moment if the `root` is not specified then the user will gain
access to the root of the storage back-end.