Auth-type now a mandatory parameter (breaking change)
Before, --auth-type had a default option of anonymous.
That way unftp doesn't require any arguments to start up.
But we figured this is a potential security hazard:
Forgetting to set --auth-type, or making a type like --authtype would
be disastrous.

So this is now disabled.
If you do not pass the --auth-type argument, unFTP will refuse to
start up.

If you intentionally want to run a passwordless FTP server, you'd have
to pass `--auth-type anonymous` explicitly.